requirements.txt
# HTTP
requests>=2.31.0

# RSS parsing
feedparser==6.0.11

# Config
pyyaml>=6.0

# Journal digest scraper
pandas>=2.0
sentence-transformers>=2.7
scikit-learn>=1.4
